LazyLoadedCategoryDetail renamed to CategoryWithLazyLoadedVisibleChildren to make the purpose of the class clear
- CategoryDetailFactory was removed
- CategoryDetailFactory::createLazyLoadedDetails() replaced with new CategoryWithLazyLoadedVisibleChildrenFactory::createCategoriesWithLazyLoadedVisibleChildren()
- CategoryFacade::getVisibleLazyLoadedCategoryDetailsForParent() renamed to getCategoriesWithLazyLoadedVisibleChildrenForParent()
- twig template for rendering category tree on FE now uses instances of CategoryWithLazyLoadedVisibleChildren class
